legal-app/legal-app/LegalMind_python/backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-docx==0.8.11
python-dotenv==1.0.0
openai==0.28.1
requests==2.31.0
python-multipart==0.0.6
streamlit==1.28.1